+++ src/main/kotlin/com/bitclave/node/repository/search/SearchRequestRepository.kt
@@ -12,9 +12,9 @@ interface SearchRequestRepository {
 
     fun deleteByIdAndOwner(id: Long, owner: String): Long
 
-    fun deleteByOwner(owner: String): Long
+    fun deleteByOwner(owner: String): Int
 
-    fun deleteByIdIn(ids: List<Long>): Long
+    fun deleteByIdIn(ids: List<Long>): Int
 
     fun findById(id: Long): SearchRequest?
